Card.js was used the same in both folders. Moved it as a component folder within es6
