--- client/src/components/lesson/es6/variables/Card.js
@@ -1,18 +1,0 @@
-import React from "react";
-
-export function CardHeader(props){
-  return (
-    <div className="card-header p-2"{...props}>
-      {props.children}
-    </div>
-  )
-}
-
-export function CardBody(props){
-  return (
-    <div className="card-body p-2">
-      {props.children}
-
-    </div>
-  )
-}
